Redo login window

+++ src/main/java/cz/muni/fi/pv168/project/business/model/User.java
@@ -2,9 +2,9 @@ package cz.muni.fi.pv168.project.business.model;
 
 public class User {
     private final String username;
-    private final Long Id;
+    private final Integer Id;
 
-    public User(String username, Long id) {
+    public User(String username, Integer id) {
         this.username = username;
         this.Id = id;
     }
